Move LoadingCover, classes, to own files
Moved LoadingCover to components/common and the classes helper function to
components/util.js
