+++ src/components/EmailTemplatesPage/EmailTemplatesPage.css
@@ -12,14 +12,3 @@
   z-index: 10;
   background: rgba(250, 250, 250, 0.6);
 }
-
-.loading-cover {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-}
-
-.loading-cover__indicator {
-  width: 100%;
-}
