+++ src/components/common/LoadingCover.css
@@ -1,0 +1,10 @@
+.loading-cover {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+}
+
+.loading-cover__indicator {
+  width: 100%;
+}
